MemoryPoolIterator byref structs
